requirements.txt
tensorflow==2.19.0
streamlit==1.48.1
pyngrok==7.3.0
ngrok==1.5.1
shap==0.48.0
pandas==2.2.2
numpy==2.0.2
matplotlib==3.10.0
scikit-learn==1.6.1
Pillow==11.3.0